# Dramatically reduce allocations