--- a/app/controllers/execute_controller.rb
+++ b/app/controllers/execute_controller.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 class ExecuteController < ApplicationController
   def post
     data = HTTP.post(URI.join(ENV["CODERUNTIME_HOST"], "/run").to_s, json: {snippet: params[:snippet], version: ruby_version})