--- a/app/graphql/ruby_api_schema.rb
+++ b/app/graphql/ruby_api_schema.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 class RubyApiSchema < GraphQL::Schema
   query Types::QueryType
 end