--- a/app/graphql/types/base_union.rb
+++ b/app/graphql/types/base_union.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Types
   class BaseUnion < GraphQL::Schema::Union
   end