--- a/app/graphql/types/ruby_method_type.rb
+++ b/app/graphql/types/ruby_method_type.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Types
   class RubyMethodType < Types::BaseObject
     implements SearchResultType