--- a/app/graphql/types/ruby_object_type.rb
+++ b/app/graphql/types/ruby_object_type.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Types
   class RubyObjectType < Types::BaseObject
     implements SearchResultType