--- a/app/graphql/types/search_result_type.rb
+++ b/app/graphql/types/search_result_type.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Types::SearchResultType
   include Types::BaseInterface