--- a/app/helpers/search_helper.rb
+++ b/app/helpers/search_helper.rb
@@ -1,11 +1,15 @@
 # frozen_string_literal: true
 
 module SearchHelper
-  def method_anchor(method)
-    # See https://github.com/ruby/rdoc/blob/c64210219ec6c0f447b4c66c2c3556cfe462993f/lib/rdoc/method_attr.rb#L294
-    method_name = CGI.escape(method.name.gsub("-", "-2D")).tr("%", "-").sub(/^-/, "")
+  def escape_method_name(method_name)
+    (@escaped_method_names ||= {})[method_name] ||= begin
+      # See https://github.com/ruby/rdoc/blob/c64210219ec6c0f447b4c66c2c3556cfe462993f/lib/rdoc/method_attr.rb#L294
+      CGI.escape(method_name.gsub("-", "-2D")).tr("%", "-").sub(/^-/, "")
+    end
+  end
 
-    "method-#{method.instance_method? ? "i" : "c"}-#{method_name}"
+  def method_anchor(method)
+    "method-#{method.instance_method? ? "i" : "c"}-#{escape_method_name(method.name)}"
   end
 
   def result_url(result, version:)