--- a/app/models/ruby_method.rb
+++ b/app/models/ruby_method.rb
@@ -4,7 +4,7 @@ class RubyMethod
   attr_reader :body
 
   def initialize(body)
-    @body = HashWithIndifferentAccess.new(body).deep_symbolize_keys
+    @body = body
   end
 
   def name