--- a/app/services/search/documentation.rb
+++ b/app/services/search/documentation.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Search
   class Results
     attr_reader :query