--- a/app/views/home/index.html.slim
+++ b/app/views/home/index.html.slim
@@ -1,3 +1,4 @@
+-# frozen_string_literal: true
 - description "Easily find and browse Ruby classes, modules and methods"
 - content_for :head do
   script type="application/ld+json"