--- a/app/views/layouts/application.html.slim
+++ b/app/views/layouts/application.html.slim
@@ -1,3 +1,4 @@
+-# frozen_string_literal: true
 doctype html
 html lang="en"
   head