--- a/app/views/objects/show.html.slim
+++ b/app/views/objects/show.html.slim
@@ -1,3 +1,4 @@
+-# frozen_string_literal: true
 - title @object.name
 - set_meta_tags canonical: object_url(object: @object.path, version: default_ruby_version) if default_ruby_version != ruby_version