--- a/app/views/search/index.html.slim
+++ b/app/views/search/index.html.slim
@@ -1,3 +1,4 @@
+-# frozen_string_literal: true
 - title search_query
 - noindex