--- a/config/initializers/inline_svg.rb
+++ b/config/initializers/inline_svg.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 InlineSvg.configure do |config|
   config.asset_finder = InlineSvg::WebpackAssetFinder
 end