--- a/config/initializers/rack_attack.rb
+++ b/config/initializers/rack_attack.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 Rack::Attack.throttle("code execution requests by ip", limit: 5, period: 30) do |req|
   if req.path == '/run' && req.post?
     req.ip