--- a/config/preload.rb
+++ b/config/preload.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 
 require_relative "environment"