--- a/lib/ruby_downloader.rb
+++ b/lib/ruby_downloader.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 class RubyDownloader
-  RUBY_HOST = "https://cache.ruby-lang.org/pub/ruby/%{base}/ruby-%{version}.tar.gz".freeze
+  RUBY_HOST = "https://cache.ruby-lang.org/pub/ruby/%{base}/ruby-%{version}.tar.gz"
   RUBY_MASTER_URL = "https://codeload.github.com/ruby/ruby/zip/master"
 
   attr_reader :version