--- a/lib/rubyapi_rdoc_generator.rb
+++ b/lib/rubyapi_rdoc_generator.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 class RubyAPIRDocGenerator
   SKIP_NAMESPACES = [
     /Bundler\:\:.*/,