--- a/lib/tasks/import.rake
+++ b/lib/tasks/import.rake
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require_relative "../ruby_downloader"
 require_relative "../ruby_documentation_importer"