--- a/test/helpers/application_helper_test.rb
+++ b/test/helpers/application_helper_test.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 class ApplicationHelperTest < ActionView::TestCase
   test "github_url" do
     method = ruby_object(String).ruby_methods.first