--- a/test/integration/search_flow_test.rb
+++ b/test/integration/search_flow_test.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require "test_helper"
 
 class SearchFlowTest < ActionDispatch::IntegrationTest