--- a/test/models/ruby_method_test.rb
+++ b/test/models/ruby_method_test.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require "test_helper"
 
 class RubyMethodTest < ActiveSupport::TestCase