--- a/test/models/ruby_version_test.rb
+++ b/test/models/ruby_version_test.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require "test_helper"
 
 class RubyVersionTest < ActiveSupport::TestCase