--- a/test/services/search/query_test.rb
+++ b/test/services/search/query_test.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require "test_helper"
 
 class Search::QueryTest < ActiveSupport::TestCase