Issue #191 points out that objects#show does _a lot_ of allocations. This PR removes about 32.5k allocations from the the String class page for Ruby 2.7 (https://rubyapi.org/2.7/o/string).

(Note: I did each optimization as a commit. The PR might be easier to review by stepping through he commits.)

# Optimization 1

## Avoid deeply nested hash and array allocations

deep_symbolize_keys returns new objects the entire way through the document tree. Additionally, the HashWithIndifferentAccess.new was an extra allocation per RubyObject and RubyMethod.

The Ruby 2.7 String class has 146 methods, which means 147 allocations just in HashWithIndifferentAccess calls.

Before:
Completed 200 OK in 72ms (Views: 57.9ms | Allocations: 58876)

After:
Completed 200 OK in 51ms (Views: 38.7ms | Allocations: 47559)

Removes ~11,000 allocations

# Optimization 2
## Add frozen_string_literal comment to all applicable files

The magic_frozen_string_literal gem made it very easy to add the "magic comment" to all Ruby files.

Before:
Completed 200 OK in 51ms (Views: 38.7ms | Allocations: 47559)

After:
Completed 200 OK in 54ms (Views: 40.2ms | Allocations: 44923)

Removes ~2,600 allocations

# Optimization 3
## Avoid `url_for` for anchor links
Rails' `url_for` method (which object_path uses) is overkill for simply creating a link to an anchor on the current page.

Each call to `object_path` allocates a Hash, and then the Rails `url_for` logic performs a lot more allocations and router lookups, just to inevitably give us `/2.7/o/string#method-i-chr` for the `/2.7/o/string` page. For an anchor, the `/2.7/o/string` is actually irrelevant. `#method-i-chr` as the href will work just fine.

Additionally, the method names are CGI escaped. A single page will escape the same method name 2-4 times, (depending on if there's both an instance and class variable with the same name). We can remove great deal of allocations by caching the escaped name for the duration of the request.

Before:
Completed 200 OK in 54ms (Views: 40.2ms | Allocations: 44923)

After:
Completed 200 OK in 38ms (Views: 24.8ms | Allocations: 26282)

Removes ~19,000 allocations

# Summary
Before:
Completed 200 OK in 72ms (Views: 57.9ms | Allocations: 58876)

After:
Completed 200 OK in 38ms (Views: 24.8ms | Allocations: 26282)

Removes ~32,600 allocations and 30ms